api: allow a start time (&start=...), end time (&end=...) and resolution(&resolution=...) to be specified
Either an interval or start parameter is required, while the end
parameter is optional when start is specified. Start and end should both
be specified in unix time format. Resolution is in seconds and should be
one of [60 | 900 | 86400 | 604800] meaning [minute | 15min | day | week]
resolution respectively.

The [time, power] pairs returned by the api will be aligned to multiples
of the specified resolution, or maximum resolution available for the
requested time interval.